Move RefreshOutcome to a leaf module and redact where the error is recorded
Round-2 review raised three things.

RefreshOutcome now lives in its own leaf module. It is the contract between
the manager and the runner, not manager internals, and the placement was
what forced the privata workaround: with the type in manager.py, privata
only accepted it as public because the runner imported it by name, and the
import only survived ruff because of an annotation that existed for no
other reason. Verified empirically -- leaf module, no annotation, no
import in the runner, privata clean. A four-line "do not simplify this
away" comment is deleted rather than kept.

Its docstring described two outcomes while the code returns three: a
non-semantic base reports published=False with no error, so tying
error is None to success was wrong. It now names published, valid no-op,
and failed, and says outright that a null error does not imply a
publication. published is kept: the terminal log line reads it, and
dropping it would re-split that summary into an outcome plus a loose flag.

Redaction moves from the outcome construction to a _record_refresh_error
helper at each assignment site. That keeps _last_refresh_error and
outcome.error identical -- they could previously differ for one run, so an
operator reading the attribute saw something other than what was logged
and persisted -- and it takes redact_credentials_in_text, which is not
total (an unparseable IPv6 URL raises ValueError), out of the finally that
owns checkpoint compaction, where a raise would mask the real failure and
skip durability work. Making the redactor total remains #1716's job.
