+++ src/mindroom/knowledge/manager.py
@@ -81,6 +81,7 @@ from mindroom.knowledge.redaction import (
     redact_credentials_in_text,
     redact_url_credentials,
 )
+from mindroom.knowledge.refresh_outcome import RefreshOutcome
 from mindroom.logging_config import get_logger
 from mindroom.strict_knowledge import StrictInsertKnowledge as Knowledge
 
@@ -290,23 +291,6 @@ class _CandidateReconciliation:
 
     expected: frozenset[str]
     pending: tuple[Path, ...]
-
-
-@dataclass(frozen=True, slots=True)
-class RefreshOutcome:
-    """What one ``KnowledgeManager.reindex_all`` call actually did.
-
-    A refresh either publishes a candidate that matches the source or explains
-    why it could not, so callers need both facts and the file count. ``error``
-    is already credential-redacted and is safe to persist or show.
-    """
-
-    #: Files this pass embedded, as opposed to reused from an earlier candidate.
-    indexed_count: int
-    #: Whether this pass swapped a complete candidate in as the published index.
-    published: bool
-    #: Why the refresh finished without publishing, or ``None`` on success.
-    error: str | None
 
 
 @dataclass
@@ -2460,6 +2444,18 @@ class KnowledgeManager:
         )
         run.journal_appends = 0
 
+    def _record_refresh_error(self, detail: str) -> None:
+        """Store why this refresh failed, redacted once at the point of record.
+
+        Redacting here rather than when the outcome is built keeps the stored
+        string and the reported one identical, so an operator reading the
+        attribute in a debugger sees exactly what was logged and persisted. It
+        also keeps ``redact_credentials_in_text`` -- which is not total -- out
+        of the ``finally`` that owns checkpoint compaction, where a raise would
+        both mask the real failure and skip durability work.
+        """
+        self._last_refresh_error = redact_credentials_in_text(detail)
+
     async def reindex_all(self, *, force_reindex: bool = False) -> RefreshOutcome:
         """Advance the durable candidate index and publish it when it matches the source.
 
@@ -2490,21 +2486,14 @@ class KnowledgeManager:
                 await self._advance_candidate(run, progress)
             except Exception as exc:
                 if self._last_refresh_error is None:
-                    self._last_refresh_error = redact_credentials_in_text(str(exc))
+                    self._record_refresh_error(str(exc))
                 raise
             finally:
                 progress.retrying = self._embedding_retry_count
-                # Redact at the seam rather than at each assignment site. Every
-                # source feeding the accumulator today is already safe --
-                # describe_embedder_error never echoes provider text, the git
-                # path redacts before raising, and the raising path below
-                # redacts -- so this pins the property for future error sources
-                # instead of cleaning up a known leak.
-                raw_error = self._last_refresh_error
                 outcome = RefreshOutcome(
                     indexed_count=progress.indexed_this_run,
                     published=run.published,
-                    error=None if raw_error is None else redact_credentials_in_text(raw_error),
+                    error=self._last_refresh_error,
                 )
                 progress.log_summary(outcome)
                 await self._finalize_candidate_checkpoint(run)
@@ -2613,7 +2602,7 @@ class KnowledgeManager:
                 summary = f"Indexed {len(run.completed)} of {len(plan.expected)} managed knowledge files"
                 if self._last_file_index_error is not None:
                     summary = f"{summary} (first error: {self._last_file_index_error})"
-                self._last_refresh_error = summary
+                self._record_refresh_error(summary)
                 return
 
             candidate_signatures = {
@@ -2622,8 +2611,8 @@ class KnowledgeManager:
                 if relative_path in expected_paths
             }
             if set(candidate_signatures) != expected_paths:
-                self._last_refresh_error = (
-                    f"Indexed signatures covered {len(candidate_signatures)} of {len(expected_paths)} managed files"
+                self._record_refresh_error(
+                    f"Indexed signatures covered {len(candidate_signatures)} of {len(expected_paths)} managed files",
                 )
                 return
 
@@ -2646,8 +2635,8 @@ class KnowledgeManager:
             await self._publish_candidate(run, candidate_source_signature)
             return
 
-        self._last_refresh_error = (
-            "Knowledge source kept changing during refresh; candidate progress was kept for the next refresh"
+        self._record_refresh_error(
+            "Knowledge source kept changing during refresh; candidate progress was kept for the next refresh",
         )
 
     async def _publish_candidate(self, run: _CandidateRun, source_signature: str) -> None:
